Fragment conversion, email search and color animations

+++ app/src/main/java/com/jamie1192/haveibeenpwned/database/models/Breach.kt
@@ -27,7 +27,7 @@ data class Breach(
         @SerializedName("ModifiedDate")
         var modifiedDate : String?,
         @SerializedName("PwnCount")
-        var pwnCount : Int?,
+        var pwnCount : Long?,
         @SerializedName("Description")
         var description : String?,
         @SerializedName("LogoPath")
